# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "recaptcha_token",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "headers": {
                            "Authorization": {
                                "description": "Token",
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Login error"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/auth/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get auth user data",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh user token",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "headers": {
                            "Authorization": {
                                "description": "Token",
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get list of news",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/News"
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "403": {
                        "description": "Have no premissions"
                    },
                    "404": {
                        "description": "User is not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "family",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "second",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "birthday",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "attachment",
                        "in": "formData",
                        "required": false,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "403": {
                        "description": "Have no premissions"
                    },
                    "404": {
                        "description": "User is not found"
                    }
                }
            }
        }
    },
    "definitions": {
        "News": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "short_description": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "User": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "family": {
                    "type": "string"
                },
                "second": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "birthday": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "remember_token": {
                    "type": "string"
                },
                "email_verified_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}